add initial state, read and mutate multimethod for counter
the read default implementation does a lookup of the key in state if
it finds it it returns the value otherwise retuns not-found.

This is useful for local keys and we can use it to fetch :tudu/count in
our case.

To increment the counter we define mutate for tudu.counter/increment
that takes a parameter called amount which is the amount to increment
the counter.

We also define our initial state as an atom with the initial counter
value using defonce so it's not redefined when figwheel reloads changes.

we pass the state and the parser (with the read and write multimethods
as parameters) to create the reconciler.

on button click we transact! passing the component as parameter (we can
also pass the reconciler here but it's easier to pass the reference to
the component) and then the query, in this case we call
tudu.counter/increment and amount set to 1, after the call we pass the
keys that we want to re read since they will change with this call.